Move the agent's protocols in with the agent
The package root declared Model, ContextSummarizer and Workspace, and had to
import the compiler to do it: Workspace.compile_path is annotated with
CompilePayload. So `import mini_articraft` -- which most modules do only for
package_dir -- pulled in the compiler, and the root package depended on a
subpackage. That was the one edge in the dependency graph running the wrong
way.

All three are the agent's swappable pieces: which model answers, how context
is compacted, where the work runs. Every implementation is already inside
agent/ and the only consumer is agent/harness.py.

They go in agent/protocols.py rather than agent/__init__.py, because
__init__ imports harness and harness needs the protocols -- via __init__ that
would cycle. agent/__init__ re-exports them, so `from mini_articraft.agent
import Model` works.

CompilePayload stays in compiler/result.py. The compiler produces the payload,
so moving the type up into the agent would invert the one edge that was
already correct.

The root is now package metadata and nothing else: `import mini_articraft`
pulls in no submodules at all, and the graph is a strict chain.
