+++ src/mini_articraft/agent/__init__.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
 from mini_articraft.agent.harness import Agent, AgentConfig
+from mini_articraft.agent.protocols import ContextSummarizer, Model, Workspace
 
-__all__ = ["Agent", "AgentConfig"]
+__all__ = ["Agent", "AgentConfig", "ContextSummarizer", "Model", "Workspace"]
